[gtk] RomDataView: Very preliminary attempt to port to GTK+ 3.x.
GTK+ 3.x won't be supported in rom-properties-0.9, but these changes
were simple enough to do and they don't really interfere with anything.

The biggest change is that GtkHBox and GtkVBox are deprecated in favor
of a single GtkBox. (GtkBox is also deprecated in favor of GtkGrid,
but GtkBox isn't generating deprecation warnings, so I'm using that
for now.)

TODO:
- Use GtkGrid instead of GtkTable.
- Use GtkWidget alignment functions instead of GtkMisc.
- Implement these in such a way that it still works with GTK+ 2.x.