[gtk] RomDataView: Make sure checkboxes can't be toggled, not that they can't be checked.
Previously, the signal handler allowed checkboxes to be unchecked but
not checked, which doesn't help if we have checkboxes that are supposed
to be checked. This patch adds an unordered_map that contains the
bitfield checkboxes and the list of checkbox states, which is used to
ensure that the checkboxes maintain their correct state.