Initial Nautilus (GTK+ 3.x) frontend.
The Nautilus frontend is very similar to the Thunar (XFCE) frontend,
though some things have changed due to the different ways Nautilus
handles things. In particular, RomPropertiesPage is not a subclass
of NautilusPropertyPage; instead, RomPropertiesPage is added to a
generic NautilusPropertyPage as a widget.

Added Find*.cmake files for glib2, GTK+ 3, GdkPixbuf2, and
libnautilus-extension.

RomDataView: Set the GtkBox orientation to vertical when compiling with
GTK+ 3.x.