[gtk] RomDataView: Use GtkGrid for the main table on GTK+ 3.x.
GtkTable is deprecated in favor of GtkGrid. GtkGrid's also a bit
easier to work with than GtkTable, since it uses x/y/w/h instead
of x1/x2/y1/y2. Also, the grid size doesn't have to be manually
adjusted. (I'm not sure if GtkTable will show items that aren't
in bounds based on the last resize, but considering resizes are
needed, it probably doesn't.)

GtkTable is still used in GTK+ 2.x builds.

lblDesc is now only created if the value widget was created
successfully, since the table attachment was moved to after
the value widget creation.

NOTE: The lblDesc changes can't be ported to the KDE version easily
because QGridLayout is a QLayout, not a QWidget.