[gtk] Notify watchers if the object properties have been changed.
GObject's documentation says this isn't needed if we use
g_object_set_property(), but it won't emit the notification a second
time if we do it ourselves. Nautilus 3.x as well as GTK+ 2.x and 3.x's
own widgets emit signals themselves, so we'll do that too.

Reference: https://developer.gnome.org/gobject/stable/gobject-The-Base-Object-Type.html