[gtk] RomDataView: Added a property "desc_format_type".
This is used to indicate how the description label is formatted.
XFCE and GNOME format it differently, so in order to fit in to the
respective DE, we have to match.

I was originally going to have separate RpDescFormatType enum values
for GTK2 vs. GTK3, but since the two versions can't be mixed, there's
no point in doing that, so just use one version for each DE.
Version-specific changes will be applied using #ifdefs later.

Sidenote: GObject enum properties are painful to deal with. I should
make use of glib-mkenums later, but CMake has issues with embedded
'@' symbols, since it uses them for variable replacement too.