[kde] RomDataView: romData is now a property.
Reworked the widget initialization functions so the widgets can
be cleared and re-created if romData is changed, similar to the
GTK+ version.

The header row widgets are now created in Ui::setupUi(). They're no
longer deleted/recreated in initHeaderRow(); instead, they're reused
and/or hidden as necessary. The "romData != nullptr" assertion has been
removed, since this condition is now allowed.

New function clearLayout(). This function clears the QFormLayout before
it's deleted, since deleting the QFormLayout doesn't delete the widgets
stored in it.

Added a constructor that only takes a parent QWidget. In this case,
romData will be nullptr, and nothing will be displayed until romData
is set manually.