[libromdata] RomData: Use reference counting.
This prevents ownership issues when using RomData objects in multiple
places. In particular, the KDE version has a RomData* property, which
is set to a RomData object that was opened elsewhere.

The GTK+ and Win32 frontends don't do this, but they might later.

RomData::~RomData is now protected. ref() and unref() must be used.
Because of this, unique_ptr<> can't be used. (Well, it can, but a
custom deleter is needed.)

TODO: Use atomic functions to manipulate the reference count.