[win32] RP_ShellPropSheetExt: Remove szSelectedFile.
It's not used anywhere except for Initialize(), which is when it gets
the filename using DragQueryFile(). It's also limited to MAX_PATH,
which prevents use of long filenames.

The function has been rewritten to use a local buffer and now uses
dynamic memory allocation, which allows it to support files longer
than 260 characters. (Whether or not Windows Explorer supports it
is another question.)

This was a leftover from the CppShellExtPropSheetHandler example:
https://code.msdn.microsoft.com/windowsapps/CppShellExtPropSheetHandler-d93b49b7