[rpcli] Use gmtime_s() and localtime_s() if they're available.
Windows has these functions, which are similar to the _r() functions but
with swapped parameters.

CMakeLists.txt: Copied CHECK_FUNCTION_EXISTS_OR_INLINE() from Gens/GS II.
This is needed because the _s() functions are macros that are defined
to the appropriate 32/64 function depending on configuration on newer
versions of MSVC (2005+). This also works for MinGW, which has its own
_r() wrappers that wrap around the _s() functions.