[rpcli] properties.cpp: Optimize JSON text escaping.
Process the string for escape characters once, instead of processing
it once per character to replace. Also eliminated unnecessary string
copies caused by passing the string by value and then returning a copy
of the string.

Sidenote: string::replace() operates on the string in question.
Its return value is a reference to the same string. Hence, returning
a copy of the string is just a waste of cycles.