[libromdata] GameBoyAdvance: Detect expansion cartridges for DS games.
The DS game "Daigasso! Band Brothers" had an expansion that was released
as a GBA cartridge. This cartridge doesn't have valid logo data, so it
isn't bootable, and previously wasn't detected.

GameBoyAdvance now checks some other fields if the logo data isn't there.
In particular, it checks if 0x96 is present as well as the device type.
If they are correct, it verifies the header checksum. If the checksum
is valid, then it assumes the ROM is a GBA ROM.

TODO: Indicate that the cartridge is not bootable.