[libromdata] GameBoyAdvance: Indicate if the ROM is an NDS expansion.
GameBoyAdvance now has a RomType enum, which currently indicates if the
ROM image is a standard GBA ROM or an NDS expansion.

In the latter case, the "Entry Point" is displayed as "Not bootable",
since the ROM is only usable as an expansion pack for a supported
Nintendo DS game.