[libromdata] GameBoyAdvance: Handle unlicensed pass-through cartridges.
These cartridges don't have Nintendo logo data, but do have a valid
header checksum and entry point. (The entry point for NDS expansions
is 0xFFFFFFFF, which is now used to detect this case.)

Convert NULLs in the game ID to '_' in order to display the partial
game ID for unlicensed pass-through cartridges.