Fixed some icon animation issues with certain DSiWare titles.
The first sequence token in "The Legend of Zelda: Four Swords Adventures"
is 0x0F05. This means physical frame 0x05 is used. Due to an optimization
for non-animated icons, this ended up being used as frame 0, which
resulted in the wrong frame being used.

This was spotted after APNG export was added, since the APNG was correct.

Updated all UI frontends to handle frame 0 in animated icon sequences
correctly.

IconAnimHelper: Set last_frame_number to the first frame in the sequence
when initializing.