[libromdata] New templated class TCreateThumbnail.
This class contains most of the code that was previously duplicated in
the KDE, Windows, and now GTK+ frontends.

Note that since this class is templated, and since the template argument
is UI-specific, it can't be compiled in the libromdata project directly.
Instead, #include "TCreateThumbnail.cpp" in the relevant class in the
UI frontend.