[libromdata] RomFields: New field type RFT_AGE_RATINGS.
This field contains up to 16 age ratings. The indexes and bitfield format
are similar to Nintendo Wii and 3DS.

This allows for better parsing of the age ratings in frontend, including
the possibility of using actual age rating logos.

GameCube: Use RFT_AGE_RATINGS for the Wii age rating field.

[kde] RomDataView: Updated to support RFT_AGE_RATINGS.