[win32] RP_ExtractImage: Reworked the private class to use TCreateThumbnail.
As an added bonus, this provides integer nearest-neighbor rescaling for
images that use IMGPF_RESCALE_NEAREST.

NOTE: Windows 7 seems to fall back to IExtractIcon if IThumbnailProvider
fails, which it does in the case of .vmi files due to the stream not
providing a full pathname. This results in the small icon being upscaled
using a non-integer nearest-neighbor multiple instead of a larger icon
(that was integer upscaled using nearest-neighbor by TCreateThumbnail)
being downscaled.