[win32] RegKey::RegisterComObject(): Disable process isolation in debug builds.
This only really applies to IThumbnailProvider on Vista+, but we'll
apply it to all objects just in case.

Disabling process isolation makes it possible to actually debug the
RP_ThumbnailProvider class, since process isolation puts it into a
COM surrogate, which may appear or disappear randomly.

Note that release builds will delete the "DisableProcessIsolation" key
on registration.