[win32] RP_ExtractIcon: Initial support for "fallback" handlers.
If a file type has a DefaultIcon and/or IconHandler registered already,
it will be saved in the RP_Fallback key. If the file type has a ProgID
that has a DefaultIcon and/or IconHandler registered, that will be saved
in RP_Fallback, too. The ProgID will also be updated.

DllMain.cpp: Temporarily disabled all extensions except for ".iso".

TODO:
- Restore RP_Fallback when unregistering.
- Use RP_Fallback if RomData can't identify a file.