[win32] RP_ExtractIcon_Fallback: Implemented IconHandler fallbacks.
The IconHandler COM object is instantiated and then processed just like
our own Iconhandler.

Tested with mhtmlfile. (Note that the mhtmlfile icon handler refuses to
check files that don't have a ".mht" or ".mhtml" extension.)

NOTE: If GetIconLocation() doesn't return GIL_NOTFILENAME, then we have
to use the returned filename and icon index as if it was a regular
DefaultIcon entry.

stdafx.h: #include <comdef.h> for _com_ptr_t.