[win32] Use hkcr+ext for RP_ExtractImage and RP_ThumbnailProvider.
- Use recursive key creation where possible to reduce API calls.
- Fixed a potential memory leak in DllMain.cpp::RegisterFileType().
  (pHkey_fileType might not be deleted if unsetting the old ProgID
  failed for some reason.)

TODO: Implement fallbacks for both IExtractImage and IThumbnailProvider.

NOTE: It might not be possible to use fallbacks for "Treatment", since
Windows Explorer reads this value itself, but we can at least save it
and restore later. (Will need to implement a way to tell the difference
between a value of 0 and no value.)