[win32] RP_ExtractImage_Fallback: Implemented IExtractImage fallbacks.
dwRecClrDepth and dwFlags are now saved in GetLocation(), since they're
needed in order to use the fallback function.

NOTE: This works for .mpg on XP, but .jxr on Win7 seems to be ignoring
the IExtractImage handler. Granted, this shouldn't be a major problem,
since IThumbnailProvider superceded IExtractImage starting with Vista...