rpcli: escaped control codes in SafeString
fixes #38

I've decided to use unicode characters (and not escape sequences)
for control codes, as normal output is supposed to be "human readable",
whereas json output is machine readable. Looks ugly on Windows, so maybe
it's not a good idea.

The only two places where the escaping was added are StringField
output (already using SafeString) and ListDataField. I don't think
it's sensible to add escaping to other places - if you have a newline
in a field name, you're pretty much asking for trouble anyway.

In ListDataField, no changes are required to the column width algorithm,
because the replacement is done one-to-one, i.e. any control code is
counted as one charcter, and any corresponding escape sequence is
logically a single character (physically, it's 3 chars, because UTF-8).

Also, removed an outdated TODO in properties.cpp :^)