[libromdata] TCreateThumbnail: Return the image size in getInternalImage() and getExternalImage().
getImgSize() on Windows can be an expensive operatino, especially with
icons, since it has to make copies of both the icon and the icon mask.

If we return the size in the functions that get the rp_image* in the
first place, then we won't have to retrieve the image data from the
OS and/or UI toolkit.

Removed getImgSize() from all subclasses.