[rpcli] properties.cpp: Handle the UTF-8 string as unsigned char when checking for control codes.
char is signed on x86, which resulted in the "*str < 0x20" condition
being true for various UTF-8 sequences. This resulted in the "©"
symbol in D43P01 (Zelda Master Quest for GameCube) being turned into
mojibake.

Also clean up some of the control character escape code.

This bug was reported by @einstein95 in #tcrf.