[extlib] libpng, zlib: Set the Windows subsystem.
The subsystem was being set to 6.0 in 64-bit builds, which prevents
usage on XP64 and Server 2003 64-bit. Use SET_WINDOWS_SUBSYSTEM()
to fix this.

Also mark the DLLs as using the Windows subsystem. (Not that it matters,
since that field is only relevant for EXEs...)

TODO: Check other DLLs before committing, and do this before the
/TSAWARE commit.