[libromdata] RomFields: New Field struct for the dynamic field setup.
RomData still uses the old setup for initialization. It's copied into
the dynamic Field vector for now.

The addData*() functions have been updated to use the new Field vector.

NOTE: I decided not to keep compatibility with desc() and data(), since
that would be a ton of extra code and memory management that would be
removed once the conversion is done. As such, the UI frontends will need
to be updated in order to compile and run correctly.

NOTE 2: Age Ratings is now stored in an std::array<> instead of a generic
C array. There's no overhead, but it requires some adjustments in various
places.