[kde] RomDataView: Updated to use the new RomFields::Field struct.
RomFields.hpp: Copied the StringFormat enum out of StringDesc. It's used
by the various UI frontends, but StringDesc will be removed once all of
the RomData subclasses have been converted to the dynamic field system,
so we need to keep a copy of it.

FIXME: Nintendo DS ROMs are showing duplicated text in the "Full Title" field.