[libromdata] NintendoDS: Use ARRAY_SIZE() for title lengths, not sizeof().
The titles are encoded using UTF-16LE, so sizeof() returned a value that
was twice the actual length in characters. This didn't cause a problem
before, since the OLD allocation method used C strings, and the title
fields are NULL-padded. The NEW allocation mtehod uses rp_string, which
uses an explicit length, so the NULLs are ignored and the next language's
title (which is sometimes the same as the requested one) is included as
part of the full title.

This *may* have caused an out-of-bounds read if the requested language
was Korean (last entry), but the worst that would happen is a crash.

This fixes the KDE frontend when using the NEW RomFields::field()
function to retrieve the field description and data.