[libromdata] RomFields: New addField*() functions.
These functions use dynamic memory management instead of the
statically-allocated versions. This makes it easier to support
multiple variants of a similar file within the same RomData
subclass, and it will allow for easier implementations of
translations later on.

isDataLoaded(): No longer deprecated; this has been updated to support
both the OLD and NEW methods.

Added assertions to prevent mixing of OLD and NEW methods. dataCount
is now set to -1 if initialized using the NEW method.

Added a no-argument constructor, which will be used with the NEW method.

RomData: Added constructors that don't take statically-allocated field
parameters.