[win32] RP_ShellPropSheetExt: Updated to use the new RomFields::Field struct.
Basically the same as the KDE changes, but modified for Windows.

RFT_LISTDATA: The ListData pointer was being set as the ListView's
window text in CreateWindow(). This didn't seem to cause any problems,
but it's wrong, since that isn't a string. Use nullptr instead.

All of the UI frontends have been updated, as well as a few RomData
subclasses. The rest of the RomData subclasses will need to be updated
before merging to master.