[libromdata] IconAnimData: Use std::array<>.
Updated the KDE and Win32 frontends to use std::array<> for the
UI-specific image objects.

NOTE: The GTK+ frontend has NOT been updated to use std::array<>, since
the GObject-derived struct is initialized by C code instead of C++.
std::array<> *is* an aggregate type, though, so it might work anyway.
Need to test this later.