[libromdata] PEResourceReader: New function loadResDir().
This loads a resource directory at the specified address.

Note that instead of reading a vector of IMAGE_RESOURCE_DIRECTORY,
it reads a vector with a custom struct that contains the resource
ID and the address of the IMAGE_RESOURCE_(DIRECTORY|DATA_ENTRY)
struct of the resource.

NOTE: The high bit of the address indicates if the entry is a subdirectory
(if set) or an actual resource (if clear).

Removed map_res_types, since there's a maximum of 24 resource types.
A linear search shouldn't be too slow.

TODO: Add convenience functions to load images and the version resource
in addition to implementing the generic open() function.