[libromdata] NintendoDS: Check the NDS Secure Area Type.
Based on ndstool:
- https://github.com/devkitPro/ndstool/blob/master/source/header.cpp#L39

"The Legend of Zelda: Four Swords Anniversary" (DSiWare) seems to have
random data in the 0x1000-0x3FFF, which causes it to be detected as
"Mask ROM". I suppose that's technically correct, since this area is
normally undumpable for cartridges...

TODO: Verify the Secure Area CRC. This requires re-encrypting the
Secure Area for decrypted ROMs, which requires the Blowfish key
and reimplementing the Secure Area algorithm.