[libromdata] Added ".srl" for GameBoyAdvance and NintendoDS.
This is the file extension used by the official SDK. The format is
identical to .gba and/or .nds, though Nintendo DS .srl files usually
have an encrypted Secure Area.

This fixes #49. (Another file extension for GBA/NDS)

README.md: Added ".srl" and noted that NDS supports "plain binary"
instead of just "decrypted". (The only encrypted portion of the ROM
is the Secure Area, which isn't used by NintendoDS except to check
the Secure Area type.)