[libromdata] PEResourceReader: Implemented open().
Added unordered_map caches for the top-level (type) and second-level (ID)
resource subdirectories.

startOfResourceSection is *not* needed; all addresses are relative to the
beginning of .rsrc, except for the actual data, which is an RVA.

Added rsrc_va to the constructors.

New functions getTypeDir() and getTypeIdDir(), which get the type and/or
type+id directory from the cache, or from .rsrc if they aren't cached yet.

EXE.cpp: Attempt to load VS_VERSION_INFO. It seems to work, but I'm not
parsing it yet.