[kde] RomDataView: Added support for subtabs.
Each tab has its own QVBoxLayout and QFormLayout, plus a QLabel
for the credits row if present.

STRF_CREDITS is now supported per-tab instead of per-instance.

Removed hasStrfCredits. This is now handled by checking if lblCredits
is nullptr.

NOTE: If only a single tab is present, a QTabWidget isn't created.
The ui.tabs vector is instead populated with a "fake" tab consisting
of a new QFormLayout, and its QVBoxLayout points to the main one.
Hence, we need to check if tab.vboxLayout == ui.vboxLayout before
deleting it.