[win32] RP_ShellPropSheetExt: Preliminary support for subtabs.
Windows doesn't support layouts like GTK+ and KDE, so I'm implementing
subtabs in individual pieces. This commit creates the WC_TABCONTROL
and adds tabs if the RomData subclass has tabs. Everything's still
drawn as if it's one tab, and switching tabs does nothing.

Reworked some of the size calculation to be more flexible.
Use InflateRect() where useful instead of open-coding.

NOTE: The inner margins for the tab are half of the outer margins.
The full margins wasted too much space.