[libromdata] RomData: Removed BOX_FULL; renamed BOX to COVER; added new BOX.
"Box" is used on GameTDB's DS section for a cover scan with the outer
box, whereas "Cover3D" is used in the Wii/GCN section for a cover scan
with an angled box rendering.

"Cover" is just for the front cover.

Removed the full cover option, since it doesn't seem like a likely choice
for thumbnailing. (I may add it back later if requested.)