[libromdata] RomData::ImageSizeDef: Added an "index" field.
This will be used for EXEs in order to improve lookup once the size
is sorted. This way, we'll just have to look it up by index instead
of by size.

Added 0s for the field everywhere except for NintendoDS, which has
the index set in ascending order for IMG_EXT_COVER. The field isn't
being used though, since the name is used for IMG_EXT_COVER instead.

TODO: Overload the name field and require checking for <= 0xFFFF,
like Windows resource IDs? (I'd rather not...)