[libromdata] NintendoDS: Disable the "coverS" image type.
There's no point in downloading the "S" size here. It's useful for
devices with low-resolution screens, e.g. 3DS, but not so much for
regular PCs.

The "S" size is also in PNG format, so we'd have to add extra logic
to handle that correctly.

Note that "S" is still present in the code, but commented out.