[libromdata] RomData::selectBestSize(): New function to select the "best" size for an image.
If the vector of ImageSizeDef has only one entry, that entry is returned.
If it's empty, a nullptr is returned. Otherwise, one of the following
algorithms is used:

- If IMAGE_SIZE_DEFAULT: Return entry 0.
- If IMAGE_SIZE_SMALLEST: Find the image with the smallest dimensions.
- If IMAGE_SIZE_LARGEST: Find the image with the largest dimensions.

Added an enum value IMAGE_SIZE_MIN_VALUE, which is the lowest "special"
value allowed for the various functions.

If a size larger than 0 is specified, the function will pick the image
closest to the specified size. The image may be slightly larger.

TODO: Check width and height separately? Currently, they're handled
as the same value.

GameCube, NintendoDS: Use selectBestSize().
Amiibo doesn't use this function yet because only one size is available.
(Granted, the same is true about GameCube, but supposedly GameTDB may
add larger sizes in the future.)

ExtURL: Added a "high_res" field. This field is used to determine if an
image should be downloaded if the user disabled high-resolution image
downloads. Note that if the image already exists, it will be used
regardless of the "high_res" field.

NintendoDS: Set "high_res" on coverM and coverHQ.

TCreateThumbnail: Renamed "max_size" to "req_size"; make use of it when
calling extURLs() functions.