[libromdata] WiiU: Added GameTDB image downloads.
FIXME: GameTDB uses ID6. The plaintext disc header doesn't have the
publisher ID, which makes it impossible to get the ID6. Currently, we're
hardcoding a publisher ID of '01' (Nintendo), which works for first-party
titles, but obviously won't work for anything else.

Since the Wii U is effectively "dead" at this point, I'm going to add a
list mapping known game IDs to publishers, which will be used for the
ID6 lookup on GameTDB.