[libromdata] FileSystem: Use atomic built-ins to guard directory initialization.
I was hitting a crash when opening a directory with multiple amiibo NFC
dumps on Windows 7. The crash was occurring because one of the
configuration directory strings was being modified while it was being
read. Turns out Windows 7 was thumbnailing multiple NFC dumps using the
same COM Surrogate.

The new setup has a single initConfigDirectories() function. This function
is called by the public getCacheDirectory() and getConfigDirectory()
functions. An internal counter variable is incremented using an atomic
function. If the new counter is 0 (initial value is -1), then it proceeds
to initialize the directories if they haven't been initialized yet.
Otherwise, it waits in a usleep() (or Sleep()) loop until the directories
have been initialized.

Some other alternatives:
- Use a static C++ initializer. This is *not* guaranteed to be thread-safe.
- Initialize in DllMain() or ELF constructors.
- DllMain() runs in a lock, so if another DLL is loaded while it's
  running, the program will lock up.
- ELF constructors run before C++ constructors, so extra care is needed.

I think the atomics method is the best option for now.

TODO: Use this for KeyManager and (later) user configuration?

NOTE: I originally used InterlockedOr() to test the bool value on Windows,
but it's not available in Visual Studio 2010. I'm now using the
InterlockedExchangeAdd() function with a value of 0, which should be
equivalent.