[win32] RP_ShellPropSheetExt: Use WS_EX_TRANSPARENT on all controls.
This fixes the background color issue. Note that this style might only
work properly on Windows 2000 and later, but that's not an issue:
1. Minimum supported OS is Windows XP.
2. It's only needed with XP theming. Win2000 and earlier use the same
   color for tab backgrounds and dialog backgrounds.

Interestingly, the empty SubtabDlgProc() is still needed for proper
control transparency, at least on Windows 7. If the dialog procedure
is removed, then child widgets have the wrong background color, but
empty areas of the child dialog have the correct background color.