[win32] RP_ShellPropSheetExt: Don't add margins to the child dialog; adjust the starting position instead.
This fixes an issue with the background gradient not lining up on
Windows XP when using XP theming.

Also, the issue with subtabs using MS Sans Serif on XP was not actually
caused by subtabs. It's caused by one of the default EXE tabs (provided
by Windows) not using DS_SHELLFONT, so I can't do anything about that.