[win32] RP_ShellPropSheetExt: Reduced the outer margins of the subtab widget.
The left margin is now 9px, and the bottom margin is now 12px.

Improved right margin sizing when using subtabs a bit for RFT_LISTDATA
and others. It's now 7px on WinXP and Win7, which is close to 9px.

TODO: GetDialogRect() returns 341x405 on Windows 7. WinSpy++ reports
the dialog as being 341x408, which explains the 3px difference. Not
sure if this is always 3px or if it's font, OS, or DPI dependent.