[win32] RP_ShellPropSheetExt: Adjust the WC_EDIT size to compensate for its margins.
With the Tahoma font, the text ended up being indented by a few pixels.
Adjust the size based on the margins returned from EM_GETMARGINS in
order to align the text correctly.

WC_EDIT widgets are only clickable within the original margins, so we
have to adjust the control size instead of using EM_SETMARGINS.

Set WS_CLIPSIBLINGS on WC_EDIT controls as well as the description
WC_STATIC controls in order to prevent drawing problems.