[libromdata] RpFile_win32.cpp: Allocate the sector buffer in readUsingBlocks() temporarily.
I tried implementing a sector_num feature to use the sector buffer as a
cache, but it didn't work for some reason.

We don't need to keep the sector buffer allocated if we aren't reading
anything, so allocate it using unique_ptr<> in readUsingBlocks().

This isn't as "efficient" due to extra allocations, but I don't expect
the block device code path to be used too often.