[libromdata] Nintendo3DS: Load the entire SMDH section in loadSMDH().
There's no point in splitting header and icon load into separate
functions, especially since that results in massive code duplication
for various encrypted formats.

The SMDH section is only 14,016 bytes, so it's not like it's that
memory-intensive, anyway.

Note that icon decoding is still only done in loadIcon(), so the
loadSMDH() function won't waste time decoding the icon if it isn't
needed.