[libromdata] Nintendo3DS: Load the ticket and display the issuer.
The issuer is shown in the TMD tab, but the issuers for both the ticket
and the TMD should be the same (retail/retail, debug/debug). The actual
issuer names are different, though. We're only checking the ticket
issuer, since that's required for CIA decryption.

N3DS_Ticket_TitleKey_KeyY: New enum to identify the key required for
CIA decryption.