[libromdata] NCCHReader: New function loadKeyNormal().
This function attempts to load and verify the specified Normal key.
If not found, it will try to load and verify the specified KeyX and
KeyY, and then scramble them using CtrKeyScrambler to get the
Normal key.

This function is implemented in NCCHReader_VerifyKeys.cpp.

NCCHReader's constructor now uses loadKeyNormal() to load the CIA
title key decryption key.