[libromdata] KeyManager: Don't initialize in the constructor.
Ubuntu 14.04 initializes KeyManager on startup, even if it isn't
actually needed. This caused CtrKeyScramblerTest to fail, since it
was trying to use a C++ string that wasn't initialized yet.

Use an atomic initialization, similar to FileSystem.

Split the atomic macros from the FileSystem code to RpAtomics.h.
Changed ATOMIC_ADD_FETCH() to ATOMIC_INC_FETCH(), since Windows'
InterlockedExchangeAdd() returns the original value, not the new
value, but InterlockedIncrement() returns the new value.

NOTE: InterlockedOr() returns the old value, but that isn't a problem,
since it's run in a loop until the value changes.

This fixes a segmentation fault in CtrKeyScramblerTest on Ubuntu 14.04,
which broke the travis-ci build.

Windows: RpAtomics.h's macros are static inline functions. This is
needed in order to preserve type safety when using Windows' atomic
functions, since they take LONG, not int.