[libromdata] KeyManager: Store key names if the key couldn't be parsed correctly.
The parsing error is stored in mapInvalidKeyNames.

get(): If the key couldn't be found in mapKeyNames, check if it's in
mapInvalidKeyNames. if it is, the key was found but couldn't be parsed,
so return that error. Otherwise, return VERIFY_KEY_NOT_FOUND.

This allows for more comprehensive error reporting to the user if there's
something wrong with the key format in keys.conf.