[libromdata] Nintendo3DS: Show a warning if a required encryption key was not found.
This is checked for CCI, CIA, and NCCH.

NCCHReader: Store the KeyManager::VerifyResult from loadNCCHKeys().
This is initially set to VERIFY_UNKNOWN, and is set to VERIFY_OK if the
keys were loaded successfully. It's also automatically set to VERIFY_OK
if NoCrypto is set.

TODO: Determine which key couldn't be loaded.