[libromdata] AesCAPI: Don't use a static HCRYPTPROV.
In the base case, we'd still end up acquiring the crypto context and
destroying it every time we need to decrypt data.

In the extreme case (multiple threads trying to use it), we'd end up
with a race condition.

The extreme case is unlikely on systems that have AesCAPI but not
AesCAPI NG, so instead of using a mutex or similar to fix the race
condition, make the HCRYPTPROV non-static and acquire it for each
instance.