[win32] ConfigDialog: Display the XP thumbnail cache dialog on XP.
This dialog is displayed if the "Thumbnail Cache" cleaner isn't found
in the registry. (Better to check for functionality than to do an
OS version check.)

Set EXT_POINTER_PROP to point to ConfigDialogPrivate, which currently
only contains "bool isVista".

Some widgets need to be hidden in order to show the WC_LISTVIEW
correctly on XP. These widgets will eventually be toggled using
the option buttons for drive vs. path.

resource.rc: Added WS_CLIENTEDGE to the WC_LISTVIEW control.