[libromdata] Don't check the conf timestamps if we last checked them less than 2 seconds ago.
Minor optimization that may reduce disk access. (Or not, since the
inode would be cached by the OS.)

NOTE: This is using seconds, not nanoseconds, so the actual timing
may be off by ±1 second.

KeyManager: Moved the data objects after the configuration values,
similar to Config.

TODO: Combine common code in Crypto and KeyManager. Maybe make a base
class, ConfReader, that handles the loading and reloading functionality.