[libromdata] Config: Special ImageTypes entry "no" to disable thumbnails.
If this is the first entry in the comma-separated value, thumbnails
will be disabled for the system in question. This is indicated by a
value of RomData::IMG_DISABLED (255) being added to the vector.

NOTE: No entry or a blank entry is interpreted as "use defaults",
which is the same ordering for all systems.