[libromdata] TCreateThumbnail: Use the user's image type priority configuration.
If set, the user's configuration overrides the default, which allows
customization of which image types to use.

Config::getImgTypePrio(): New function to retrieve the image type
priority configuration. Based on KeyManager::get().

If the configuration for the specified class is empty or missing,
the function will return IMGTR_USE_DEFAULTS, which instructs
TCreateThumbnail to use the default priorities.

If the configuration is explicitly disabled using "No", the function
will return IMGTR_DISABLED, which instructs TCreateThumbnail to
exit immediately.

If any other error occurs, the appropriate error is returned.
TCreateThumbnail currently handles these cases as IMGTR_USE_DEFAULTS.

If the configuration is loaded successfully, the function returns
IMGTR_SUCCESS, which tells TCreateThumbnail to use the returned
priority list. (Internal icons will take precedence if the requested
image size is "small" and useIntIconForSmallSizes is set.)

TODO: Rewrite RP_ExtractIcon to use TCreateThumbnail.