[win32] ConfigDialog: Allow the user to save changes in the Downloads tab.
Changing any of the checkboxes sends a PSM_CHANGED message to the
property sheet, which enables the Apply button.

If either OK or Apply are clicked, save_Downloads() is called,
which uses WritePrivateProfileString() to update the INI file.

NOTE: EXT_POINTER_PROP is now D_PTR_PROP, and it contains a pointer
to the ConfigDialogPrivate object.