[libromdata] GameCube: Added partial support for WIA disc images.
A copy of the disc header is stored in plaintext, but the actual disc
contents are compressed using LZMA. I added support for the header
fields only, which means most properties are displayed and external
images can be downloaded.

Fields not displayed:
- Region code
- Age ratings

The GCN internal banner cannot be displayed at the moment.

Reference: Issue #19 (GCN/Wii: Add support for more disc image formats.)