[rpcli] properties.cpp: Don't say an internal image exists if it doesn't.
imgbf might be set because the ROM/disc image *could* contain that
image, but it's possible that a ROM or disc might be missing said
image, or it can't be extracted at the moment for various reasons.

Check if the image is actually extractable first. If it is, show the
image information. If not, act like imgbf doesn't list it.