[win32] RP_ExtractImage, RP_ThumbnailProvider: Also use rp_image::square().
If we only use this in RP_ExtractIcon, we end up with shifting when
refreshing the view in Explorer on Windows 7. Squaring the previews
for all of them fixes this issue.

NOTE: Explorer seems to use slightly different scaling methods for
icons and thumbnails, so the images are still a bit different even
with this fix.