[win32] Moved monospaced font enumeration from RP_ShellPropSheetExt to WinUI.
RP_ShellPropSheetExt no longer has to maintain an unordered_set<wstring>
containing font names. This is usually cleared after the initial search,
but it still used up memory.

Granted, this could have been fixed by passing a locally-allocated
unordered_set<wstring> as LPARAM, but moving it to WinUI reduces the
amount of clutter in RP_ShellPropSheetExt, plus it allows other classes
to make use of it.