[win32] ConfigDialog: Initial commit for splitting each tab into a separate class.
The single ConfigDialog class was getting too big, so I'm splitting
out each tab into a new class derived from a simple base class, ITab.

ITab::getHPropSheetPage() can be called once in order to get an
HPROPSHEETPAGE for the tab. If called a second time, it will assert()
and/or return nullptr.

NOTE: I may or may not do this for the Qt and GTK+ frontends, since
they usually allow a single dialog resource for all tabs instead of
requiring each tab to be a separate dialog.