[win32] ImageTypesTab: Initial save() implementation.
If a system is using the default configuration, sysIsDefault[sys] is set.
If any value for this system is changed by the user, it's cleared.

If sysIsDefault[sys] is set when the user saves, a blank string will be
written for that class.

NOTE: If no system configurations are changed, then nothing will be
written, since d->changed will be false.

TODO: Save modified configurations.