[win32] ImageTypesTab: Prevent duplicates when the user changes an image type priority.
Suppose we have:
- ImageA: 1
- ImageB: 2
- ImageC: 3
- ImageD: No

If the user sets ImageA to 3, ImageC will be switched to 1.
If the user sets ImageD to 2, ImageB will be switched to No.
If the user sets any image to No, nothing else will change.

NOTE: If there's no image with priority 1, but there are images with
priorities 2, 3, etc., closing and reopening will result in the highest
priority image being renumbered as 1. This is because the priorities are
stored as a comma-separated string of image types. The numeric priority
values are only present in the UI and in the Config class.