[win32] ConfigDialog: Set the dialog icon.
We're using a ROM chip icon from imageres.dll (Vista+) or
shell32.dll (older versions).

NOTE: PropertySheet()'s pszIcon parameter only handles the small icon,
so we have to set the large icon manually using CallbackProc.

NOTE 2: The DLL isn't freed once the icon is obtained. This is due to
the callback function not providing any way to get a parameter pointing
to the private class. This shouldn't be a problem, since ConfigDialog
is only ever used as a standalone program.

RpWin32_sdk.h: Enable system metrics. This is needed for
GetSystemMetrics() and related macros.