[win32] PropSheetIcon: New static class to load the property sheet icon.
This uses InitOnceExecuteOnce() to ensure the icon is only loaded once.

The icon is not loaded as LR_SHARED anymore. This allows us to unload
the DLL once we've loaded the icon.

NOTE: Both the small and large icon are now loaded. Previously, we
only manually loaded the large icon, since PropertySheet() could
load the icon itself given a DLL handle and resource ID.

ConfigDialog: Use PropSheetIcon to get the icons.