[libromdata] RpWin32_delayload.h: Added a DELAYLOAD check function.
This uses MSVC structured exception handling to verify that the
DELAYLOAD DLL is available. If it isn't, then it skips processing.

RpPng and RpJpeg have both been updated to use this.

EXE_manifest has also been updated to use this, but due to conflicts
with exe_structs.h and the Windows SDK, it's implemented in a separate
file, EXE_delayload.cpp. It also doesn't use the generic macros in
RpWin32_delayload.h, since TinyXML2 doesn't export a generic C function
that we can use. (It does use RpWin32_delayload.h for the #includes,
though.)

TODO:
- Only use this if the DLL is marked as DELAYLOAD.
- MinGW support.